+++ packages/app/src/modules/options/pivotTableConfig.js
@@ -25,6 +25,7 @@ import Regression from '../../components/VisualizationOptions/Options/Regression
 import Cumulative from '../../components/VisualizationOptions/Options/Cumulative'
 import SortOrder from '../../components/VisualizationOptions/Options/SortOrder'
 import TopLimit from '../../components/VisualizationOptions/Options/TopLimit'
+import ApprovalLevel from '../../components/VisualizationOptions/Options/ApprovalLevel'
 import ShowHierarchy from '../../components/VisualizationOptions/Options/ShowHierarchy'
 import CompletedOnly from '../../components/VisualizationOptions/Options/CompletedOnly'
 
@@ -65,6 +66,7 @@ export default [
                     <AggregationType />,
                     <NumberType />,
                     <CompletedOnly />,
+                    <ApprovalLevel />,
                 ]),
             },
         ],