+++ packages/plugin/src/api/legendSets.js
@@ -1,0 +1,9 @@
+const legendSetQuery = {
+    legendSet: {
+        resource: 'legendSets',
+        id: ({ id }) => id,
+    },
+}
+
+export const apiFetchLegendSet = (dataEngine, id) =>
+    dataEngine.query(legendSetQuery, { variables: { id } })